refactor: move shared task functions to utils/task
